pyproject.toml
[build-system]
requires = ["scikit-build-core >=0.10", "nanobind >=1.3.2"]
build-backend = "scikit_build_core.build"

[project]
name = "fpng"
version = "0.0.1"
description = ""
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "KEUM Myungchul" },
]
classifiers = [
]

[project.optional-dependencies]
test = ["pytest", "numpy", "Pillow"]

[project.urls]
Homepage = "https://github.com/dofuuz/py-fpng-nb"


[tool.scikit-build]
# Protect the configuration against future changes in scikit-build-core
minimum-version = "build-system.requires"

# Setuptools-style build caching in a local directory
build-dir = "build/{wheel_tag}"

# Build stable ABI wheels for CPython 3.12+
wheel.py-api = "cp312"

[tool.cibuildwheel]
# Necessary to see build output from the actual compilation
build-verbosity = 1

# Run pytest to ensure that the package was correctly built
test-command = "pytest {project}/tests"
test-extras = ["test"]

# Don't test Python 3.8 wheels on macOS/arm64
test-skip="cp38-macosx_*:arm64"

# Needed for full C++17 support
[tool.cibuildwheel.macos.environment]
MACOSX_DEPLOYMENT_TARGET = "10.14"
